Split out WebSocketClientCmd enum to file
This commit moves the WebSocketClientCmd enum into its own file to help
keep the module organized. This is the first step towards being able to
move the type to the rest api crates.
